+++ src/main/java/frc/robot/subsystems/led/Leds.java
@@ -1,13 +1,9 @@
 package frc.robot.subsystems.led;
 
-import java.time.Duration;
-import java.time.LocalTime;
-
 import edu.wpi.first.wpilibj.AddressableLED;
 import edu.wpi.first.wpilibj.AddressableLEDBuffer;
 import edu.wpi.first.wpilibj.util.Color;
 import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
@@ -249,41 +245,5 @@ public class Leds extends SubsystemBase {
 
     public Command turnOffCommand() {
         return new InstantCommand(this::turnOff, this);
-    }
-
-    // ========================= Game Commands =================================
-
-    /**
-     * Blink the LEDs to specified Color and then return to Alliance color
-     * 
-     * @param leds  leds being set
-     * @param color color being blinked to
-     * @return Blink Command
-     */
-    public Command blinkCommand(Color color) {
-        Duration WAIT_TIME = Duration.ofMillis(500);
-        Command blinkCommand = new Command() {
-            LocalTime startTime;
-
-            @Override
-            public void initialize() {
-                setColor(color);
-                startTime = LocalTime.now();
-            }
-
-            @Override
-            public boolean isFinished() {
-                Duration timeWaited = Duration.between(startTime, LocalTime.now());
-                return timeWaited.compareTo(WAIT_TIME) >= 0;
-            }
-        };
-        blinkCommand.addRequirements(this);
-        return blinkCommand;
-    }
-
-    public Command defaultLedCommand() {
-        return Commands.run(() -> {
-            setAllianceColor();
-        }, this);
     }
 }
