+++ src/main/java/frc/robot/subsystems/shooter/Flywheel.java
@@ -28,10 +28,8 @@ import edu.wpi.first.units.Velocity;
 import edu.wpi.first.units.Voltage;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
-import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.StartEndCommand;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.Constants;
 
 public class Flywheel extends SubsystemBase {
@@ -267,18 +265,4 @@ public class Flywheel extends SubsystemBase {
     public Command reverseStopCommand() {
         return new StartEndCommand(this::reverse, this::stop, this);
     }
-
-    // ========================= Game Commands =================================
-
-    public Command defaultFlywheelCommand() {
-        return new SequentialCommandGroup(new WaitCommand(.25), stopCommand());
-    }
-
-    public Command spinUpFlywheelCommand() {
-        return new SequentialCommandGroup(
-                startCommand(),
-                new WaitCommand(1) // TODO: Tune.
-        );
-    }
-
 }
